requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
pydantic[email]
pydantic-settings
alembic